Fix test script
