+++ test/entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Bytes, BigDecimal, BigInt } from 'temp_lib/index'
 
-export function test(): void {
+export function testEntity(): void {
   let entity = new Entity()
 
   entity.setBytes('x', new Bytes(1))
