+++ test/test.js
@@ -17,49 +17,48 @@ if (!fs.existsSync('test/temp_lib/chain')) {
 
 fs.copyFileSync('chain/ethereum.ts', 'test/temp_lib/chain/ethereum.ts')
 fs.copyFileSync('index.ts', 'test/temp_lib/index.ts')
-let output_path = 'test/temp_out/test.wasm'
 
-const env = {
-  abort: function (message, fileName, lineNumber, columnNumber) {
-    console.log('aborted')
-  },
-}
+const OUTPUT_PATH = 'test/temp_out/test.wasm'
 
 try {
-  testFile('test/test.ts')
-  testFile('test/testBytes.ts')
-  testFile('test/entity.ts')
-
-  // Cleanup
-  fs.unlinkSync('test/temp_lib/index.ts')
-  fs.unlinkSync('test/temp_lib/chain/ethereum.ts')
-  fs.rmdirSync('test/temp_lib/chain')
-  fs.rmdirSync('test/temp_lib')
-  fs.unlinkSync('test/temp_out/test.wasm')
-  fs.rmdirSync('test/temp_out')
+  // TODO:testFile('test/bigint.ts')
+  testFile('test/bytes.ts')
+  // TODO: testFile('test/entity.ts')
 } catch (e) {
   process.exitCode = 1
+  throw e
+} finally {
+  // Cleanup
+  fs.unlinkSync('test/temp_out/test.wasm')
+  fs.rmdirSync('test/temp_out')
   fs.unlinkSync('test/temp_lib/index.ts')
   fs.unlinkSync('test/temp_lib/chain/ethereum.ts')
   fs.rmdirSync('test/temp_lib/chain')
   fs.rmdirSync('test/temp_lib')
-  fs.unlinkSync('test/temp_out/test.wasm')
-  fs.rmdirSync('test/temp_out')
-  throw e
 }
 
-function testFile(path) {
-  if (asc.main([path, '--lib', 'test', '--validate', '-b', output_path]) != 0) {
+function testFile (path) {
+  if (asc.main([path, '--debug', '--lib', 'test', '-b', OUTPUT_PATH]) != 0) {
     throw Error('failed to compile')
   }
-  let test_wasm = new Uint8Array(fs.readFileSync(output_path))
+
+  let test_wasm = new Uint8Array(fs.readFileSync(OUTPUT_PATH))
 
   WebAssembly.instantiate(test_wasm, {
-    env,
     index: {
-      'typeConversion.bytesToHex': function () {},
+      'typeConversion.bytesToHex': function() {},
+    },
+    env: {
+      abort: function(_msg, _file, line, column) {
+        console.error("abort called at index.ts:" + line + ":" + column)
+      },
     },
   }).then((module) => {
-    module.instance.exports.test()
+    for (const [member, f] of Object.entries(module.instance.exports)) {
+      if (typeof f === 'function' && member.startsWith('testB')) {
+        console.info(`Running "${member}"...`)
+        f.apply()
+      }
+    }
   })
 }
